requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pydantic-settings
python-dotenv
simple-salesforce
google-generativeai
mcp
